Rename FunctionLoader files to LegacyFunctionLoader
