+++ src/LegacyFunctionLoader.ts
